pdf-saas/pdf-content-extracter-microservice/requirements.txt
fastapi
uvicorn
PyMuPDF
python-multipart
python-dotenv

